% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat.plotAscatProfile}
\alias{ascat.plotAscatProfile}
\title{ascat.plotAscatProfile}
\usage{
ascat.plotAscatProfile(n1all, n2all, heteroprobes, ploidy, rho, goodnessOfFit,
  nonaberrant, y_limit = 5, nAfull, ch, lrr, bafsegmented, chrs,
  textFlag = FALSE)
}
\arguments{
\item{n1all}{copy number major allele}

\item{n2all}{copy number minor allele}

\item{heteroprobes}{probes with heterozygous germline}

\item{ploidy}{ploidy of the sample}

\item{rho}{purity of the sample}

\item{goodnessOfFit}{estimated goodness of fit}

\item{nonaberrant}{boolean flag denoting non-aberrated samples}

\item{y_limit}{Optional parameter determining the size of the y axis in the nonrounded plot and ASCAT profile. Default=5}

\item{nAfull}{copy number major allele}

\item{ch}{a list containing c vectors, where c is the number of chromosomes and every vector contains all probe numbers per chromosome}

\item{lrr}{(unsegmented) log R, in genomic sequence (all probes), with probe IDs}

\item{bafsegmented}{B Allele Frequency, segmented, in genomic sequence (only probes heterozygous in germline), with probe IDs}

\item{chrs}{a vector containing the names for the chromosomes (e.g. c(1:22,"X"))}

\item{textFlag}{Optional flag to add the positions of fragments located outside of the plotting area to the plots. Default=F}
}
\value{
plot showing the ASCAT profile of the sample
}
\description{
ascat.plotAscatProfile
}

