% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat.plotRawData}
\alias{ascat.plotRawData}
\title{ascat.plotRawData}
\usage{
ascat.plotRawData(ASCATobj)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. data structure from ascat.loadData)}
}
\value{
Produces png files showing the logR and BAF values for tumour and germline samples
}
\description{
Plots SNP array data
}

